requirements.txt
fastapi==0.115.12
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
Requests==2.32.3
torch==2.6.0
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
streamlit==1.44.1
